Use our types everywhere. No more errors!
