+++ src/stores.ts
@@ -5,64 +5,17 @@ import {
   type Writable,
   type Readable,
 } from "svelte/store";
-import type { Point, LineString, Polygon } from "geojson";
 import { emptyGeojson } from "./maplibre_helpers";
 import type { Map } from "maplibre-gl";
+import type { Scheme } from "./types";
 
 // A global singleton, containing a loaded map
 // TODO | null. When we enable strictNullChecks, this'll become a problem
 export const map: Writable<Map> = writable(null);
 
-// This describes the full structure of the GeoJSON we manage. We constrain the
-// default GeoJSON types and specify feature properties.
-export interface Scheme {
-  type: "FeatureCollection";
-  features: OurFeatureUnion[];
-  // Foreign members
-  scheme_name?: string;
-  authority?: string;
-  origin?: string;
-}
-
-// TODO Can we use a wildcard type, like OurFeature<? extends OurGeometry>
-export type OurFeatureUnion =
-  | OurFeature<Point>
-  | OurFeature<LineString>
-  | OurFeature<Polygon>;
-
-export type OurGeometry = Point | LineString | Polygon;
-
-export interface OurFeature<G extends OurGeometry> {
-  type: "Feature";
-  // Must be defined, and always > 0
-  id: number;
-  geometry: G;
-  properties: InterventionProps;
-}
-
-export interface InterventionProps {
-  intervention_type: "area" | "route" | "crossing" | "other";
-  name: string;
-  description: string;
-
-  // For LineStrings only
-  length_meters?: number;
-  waypoints?: Waypoint[];
-
-  // Temporary state, not meant to be serialized
-  editing?: boolean;
-  hide_while_editing?: boolean;
-}
-
-export interface Waypoint {
-  lon: number;
-  lat: number;
-  snapped: boolean;
-}
-
 // TODO Should we instead store a map from ID to feature?
 // TODO Plus the foreign members we add in
-export const gjScheme: Writable<Scheme> = writable(emptyGeojson());
+export const gjScheme: Writable<Scheme> = writable(emptyGeojson() as Scheme);
 
 // The optional ID of a feature currently hovered from the sidebar or map. When
 // an intervention is open and being edited, hovering is fixed to it.
@@ -79,7 +32,7 @@ export const currentlyEditing: Readable<number | null> = derived(
   ($gj) => {
     let f = $gj.features.find((f) => f.properties.editing);
     if (f) {
-      return f.id as number;
+      return f.id;
     } else {
       return null;
     }
@@ -123,7 +76,7 @@ export function setCurrentlyEditing(id: number | null) {
 // NOTE! If you call this twice in a row in a `gjScheme.update` transaction
 // without adding one of the new features, then this'll return the same ID
 // twice!
-export function newFeatureId(gj): number {
+export function newFeatureId(gj: Scheme): number {
   let ids = new Set();
   for (let f of gj.features) {
     ids.add(f.id);
